+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 15.3.0
 
 - Raise minimum supported Python version to 3.10
+- Fix [#195](https://github.com/shinolab/autd3/issues/195): missing setter for speed of sound in Geometry for C++/C#/Python
 
 # 15.2.1